Capture the Initial State of Component IssueTable
Capture the state of IssueTable component with thi.state
Use the array of issue as the state of IssueTable component
In the render method of IssueTable we change the loop the loop that creates the set of issueRows to use the state variable called issues (this.state.issues.

We still use the hard-coded set of issues that was defined in the global scope but we change the name from issues to initialIssues.

We set the initial state in the constructor of IssueTable component. We use the variable initialIssues (which is an arrau of object) to initialize the value of the state variable issues `this.state = {issues: initialIssues}`