Async State Initialization
Allow component initial state to be fecthed asynchronousy via the call to a method

Normally Component's initial state are not set statically. They are fecthed from a server mostly via an API call.
For IssueTracker application, even the initial list of issues to be displayed would have to be fetched via API call

The state can only be assign a value. But to change it we need to use the `this.setState()` method.
Here we simulate that.
At the time of component construction we don't have initial data. So we assign an empty array to the issue state variable

We the simulate the call to a server by creating a method that returns the list of issues asynchronously. The method is called loadData
Inside the method we use the async method setTimeout. The setTimeout takes a callback function which is triggered where the time fixe is out.
Here this callback function is an anonymous function within which we set issues state variable.

The loadData should not be called in the constructor method but instead in one of REACT predefine lifecycle hooks methods. The one we choose here is componentDidMount()