GraphQL Schema File
In the previous branch `graphql_about_api` we create the graphql schema definition inside the server.js file.
This can be a problem as the schema will start to grow bigger. We need to separate the schema into it's own file

We then need to change the package.json script section so that nodemon can detect a change in js files as well as
graphql files. we use the -e js,graphql in the 'npm start' script