Custom scalar type
Inside the type Issue in the schema.graphql file, we need to change the type of fields 'created' and 'due' from String to date.
But GraphQL doesn't support Date time like JS.
A workaround is to create a 'custom scalar type' of format:

...
scalar GraphQLDate
...

GraphQLDate is the name that we give to our custom scalar type

Next we need to change the String type of 'created' and 'due' field to the newly created type i.e GraphQLDate

Nex we need to create a resolver for the scalar type which will handle serialization as well as parsing
A scalar type need to be an object of the class GraphQLScalarType define in package graphql-tools

we construct the GraphQLScalarType object with an object as constructor. That object takes in properties such as name, description, but more importantly the serialize method which can convert JS date into String.

After defining the resolver, we put it at the same level as Query and Mutation

 Please enter the commit message for your changes. Lines starting