Custom Scalar type (2)
Now we need to convert the string to the native Date type.
A better way is to pass a reviver function to JSON.parse() function
A reviver function is one that is called for parsing all values
we create that reviver function that looks a date-like pattern.

We also use reponse.text() to obtain the response in text format
before passing it to JSON.parse(body, reviver)

We also revert our changes to display the date to what we had before: using toDateString()