Query Variables
In the preceding code we specify the forms field values into the query string. But since arguments are dynamic by nature because based on user input, specifying arguments directly in query string is not a good practice. GraphQL had a first-class way to factor out dynamic values out of the query and pass them as a separate dictionnary.

This is what we implemented here.