Displaying Errors
Modify the user interface to show any error messages to the user.
We'll deal with transport errors due to network connection problem
as invalid user input.

We'll a create a common utility function that handle API calls and
report errors

We replace the fetch call within the handlers with the utility
function created.